convert an unsogned int to int. ci_get_data_type_id can return -1 .....
git-svn-id: https://svn.code.sf.net/p/c-icap/code/c-icap-server/trunk/c-icap@144 950cb5d8-81e4-40ab-8cb8-c719d5138fe1

+++ filetype.c
@@ -293,7 +293,7 @@ void ci_magics_db_release(struct ci_magics_db *db)
 
 int ci_magics_db_file_add(struct ci_magics_db *db, char *filename)
 {
-     unsigned int type;
+     int type;
      int ret, group, i;
      int groups[MAX_GROUPS + 1];
      struct ci_magic_record record;
@@ -323,8 +323,7 @@ int ci_magics_db_file_add(struct ci_magics_db *db, char *filename)
                     break;
                }
           }
-
-          magics_add(db, record.offset, record.magic, record.len, type);
+          magics_add(db, record.offset, record.magic, record.len, (unsigned int)type);
           free_records_group(&record);
      }
      fclose(f);
